Show edit form (but don't persist changes) when valinnat is down
Previously, if valinnat was down and the user tried to update
phase answers, a blank form with
two save buttons was shown. This could cause data to disappear,
because if the user clicked on the save button (and meanwhile valinnat
would go up again) the save was successful and data disappeared.