OY-254 : Make most audit details mandatory
This means that we need to provide request
and user data in tests.

Previous fake logins in test used user account names
instead of OIDs in logging, producing this kind of
effect

13:09:39.322 [[ERROR] HakuAuditLogger [http-bio-9090-exec-8]] Error creating Oid-object out of officer

But it's better to have as many places as possible failing
fast if user details meant for audit log are not present.